+++ superawesome-base/src/main/java/tv/superawesome/sdk/publisher/SAVideoAd.java
@@ -580,6 +580,19 @@ public class SAVideoAd extends Activity {
         return object != null && object instanceof SAAd;
     }
 
+    public static SAAd getAd (int placementId) {
+        if (ads.containsKey(placementId)) {
+            Object object = ads.get(placementId);
+            if (object instanceof SAAd) {
+                return (SAAd) object;
+            } else {
+                return null;
+            }
+        } else {
+            return null;
+        }
+    }
+
     /**
      * Static method that, if an ad data is loaded, will play the content for the user
      *